Added code to parse arguments from envrionment variables, removed debugging rand_envp() call, added argument for blacklisting ip addresses in a file. sockets.c is still a bit of a mess. untested code.

+++ ringbuffer.c
@@ -40,7 +40,7 @@ ringbuffer_create (size_t sz)
         return NULL;
     }
 
-    for (power_of_two = 1; 1 << power_of_two < sz; power_of_two++);
+    for (power_of_two = 1; (unsigned int) (1 << power_of_two) < sz; power_of_two++);
 
     rb->size = 1 << power_of_two;
     rb->size_mask = rb->size;
